+++ CHANGELOG.md
@@ -4,7 +4,11 @@ All notable changes to Hylo Node (the Hylo server) will be documented in this fi
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## [3.3.0] - 2022-04-28
+### Added
+- Methods for adding users and groups via OAuth
+- Graphql endpoints for login and other auth-related features
+- Login by JWT route POST route without redirect for XHR requests
 
 ## [3.2.0] - 2022-03-28
 ### Added